# static/swagger.spec.json
{
  "info": {
    "title": "demo-ecommerce-backend",
    "version": "0.1.0",
    "description": "Demo eCommerce BackEnd - API supplier and Swagger Docs"
  },
  "host": "localhost:8000",
  "schemes": [
    "http",
    "https"
  ],
  "basePath": "/",
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Standard Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "swagger": "2.0",
  "paths": {
    "/api/v1/products": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Find and count all products",
        "description": "Returns a list of `ProductModel` items that matched the `select` criteria.\n",
        "operationId": "findAndCountAllProducts",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "select",
            "in": "query",
            "type": "object",
            "description": "Select",
            "default": {
              "name": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "`count` and `rows` response of `UserModel` items that matched the request parameters.\n",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "number",
                  "format": "integer",
                  "description": "Total number of records that match the `select` criteria.\n",
                  "example": 12
                },
                "rows": {
                  "description": "An array of `ProductModel` items that matched the request parameters.\n",
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/ProductModel"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/responses/AppNotFound"
          }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Find and count all Users",
        "description": "Returns a list of `UserModel` items that matched the `select` criteria.\n",
        "operationId": "findAndCountAllUsers",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "select",
            "in": "query",
            "type": "object",
            "description": "The `select` query object is used to point what data the API need to return.\nAdding a property to the `select` object will return the corresponding data.\n\nFor example, if you'd like to get a list of all `Users` with their full name, you can submit this `select`:\n```\n{\n  \"full_name\": \"\"\n}\n```\n<br />\n\nWhenever you fee like you want only a slice of the response you can apply pagination via the `$offset` and `$limit` props.\nFor example, if you'd like to get page **3** of all `Users` with a maximum of **5** rows per page, you can submit this `select`:\n```\n{\n  \"full_name\": \"\",\n  \"$offset\": 10,\n  \"$limit\": 5\n}\n```\n<br />\n\nFiltration is also supported with the special key of `\"$where\":{}`. The `$where` object supports all comparison operators\nfrom Sequelize ORM https://sequelize.org/master/manual/model-querying-basics.html#operators. Please note that operations are prefixed, e.g. instead of `Op.eq` => `$eq`.\n\nFor example, if you'd like to get a list of all `Users` within the continent of `Africa`, you can submit this `select`:\n```\n{\n  \"full_name\": \"\",\n  \"country\": {\n    \"name\": \"\",\n    \"continent_name\": \"\",\n    \"$where\": {\n      \"continent_name\": {\n        \"$eq\": \"Africa\"\n      }\n    }\n  }\n}\n```\n<br />\n\nSimilarly, you can use the `\"$order\":[]` object to set a specific order of your results. The `$order` object is an array of rules that will be applied starting from the first rule in the array down to the last one.\n\nFor example, if you'd like to get a list of `Users` which ordered the most products by country, you can submit this `select`:\n```\n{\n  \"full_name\": \"\",\n  \"country\": {\n    \"name\": \"\"\n  },\n  \"orders\": {\n    \"id\": \"\",\n    \"order_items\": {\n      \"quantity\": \"\"\n    }\n  },\n  \"$order\": [\n    [\"orders.order_items.quantity\", \"desc\"],\n    [\"country.name\", \"asc\"]\n  ]\n}\n```\n<br />\n",
            "default": {
              "full_name": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "`count` and `rows` response of `UserModel` items that matched the request parameters.\n",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "number",
                  "format": "integer",
                  "description": "Total number of records that match the `select` criteria.\n",
                  "example": 12
                },
                "rows": {
                  "description": "An array of `UserModel` items that matched the request parameters.\n",
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/UserModel"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/responses/AppNotFound"
          }
        }
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Find one User",
        "description": "This API returns a single one `User` based on common `select` request\nand the `id` URL param\n",
        "operationId": "findOneUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/PathUniqueIdentifierV4"
          },
          {
            "name": "select",
            "in": "query",
            "type": "object",
            "description": "The `select` query object is used to point what data the API need to return.\nAdding a property to the `select` object will return the corresponding data.\n\nFor example, if you'd like to get the `User` full name, you can submit this `select`:\n```\n{\n  \"full_name\": \"\"\n}\n```\n\nFiltration is also supported with the special key of `\"$where\":{}`. The `$where` object supports all comparison operators\nfrom Sequelize ORM https://sequelize.org/master/manual/model-querying-basics.html#operators. Please note that operations are prefixed, e.g. instead of `Op.eq` => `$eq`.\n\nFor example, if you'd like to get a list of all `Products` a specific `User` bought around Christmas, you can submit this `select`:\n```\n{\n  \"full_name\": \"\",\n  \"orders\": {\n    \"created_at\": \"\",\n    \"$where\": {\n      \"created_at\": {\n        \"$and\": [\n          {\"$gt\": \"2019-12-10T00:00:00.000Z\"},\n          {\"$lt\": \"2019-12-27T00:00:00.000Z\"}\n        ]\n      }\n    },\n    \"order_items\": {\n      \"quantity\": \"\",\n      \"product\": {\n        \"name\": \"\"\n      }\n    }\n  }\n}\n```\n\nSimilarly, you can use the `\"$order\":[]` object to set a specific order of your results. The `$order` object is an array of rules that will be applied starting from the first rule in the array down to the last one.\n\nFor example, if you'd like to get what was the most expensive `Order` a specific `User` paid for, you can submit this `select`:\n```\n{\n  \"full_name\": \"\",\n  \"orders\": {\n    \"id\": \"\",\n    \"order_items\": {\n      \"quantity\": \"\",\n      \"product\": {\n        \"name\": \"\",\n        \"price\": \"\"\n      }\n    }\n  },\n  \"$order\": [\n    [\"orders.order_items.quantity\", \"desc\"],\n    [\"orders.order_items.product.price\", \"desc\"]\n  ]\n}\n```\n",
            "default": {
              "full_name": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Single one `Users` that matched the request parameters.\n",
            "schema": {
              "$ref": "#/definitions/DatabaseUser"
            }
          },
          "404": {
            "description": "Unable to find User with the requested params.",
            "schema": {
              "$ref": "#/definitions/ResponseError"
            },
            "examples": {
              "application/json": {
                "errorId": "a34f4558-1b88-47cd-8d63-ae4eaf038bd7",
                "errorCode": "USER_NOT_FOUND",
                "errorMessage": "Unable to find user"
              }
            }
          },
          "500": {
            "$ref": "#/responses/AppNotFound"
          }
        }
      }
    }
  },
  "definitions": {
    "ResponseError": {
      "type": "object",
      "description": "Schema for `4XX` server error response body.\nResponse includes a UUIDv4, error code and message body.\n",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid",
          "example": "a34f4558-1b88-47cd-8d63-ae4eaf038bd7"
        },
        "code": {
          "type": "string",
          "example": "GENERIC_SERVER_ERROR"
        },
        "message": {
          "type": "string",
          "example": "Generic server error"
        }
      },
      "required": [
        "id",
        "code",
        "message"
      ]
    },
...
    "ProductModel": {
      "description": "ProductModel is the base of all that a seller can sell.\nIt has a general name, price, and props a buyer can view.\n",
      "allOf": [
        {
          "type": "object",
          "required": [
            "id",
            "name",
            "description",
            "imageUrl",
            "price",
            "basePrice",
            "props",
            "totalInStock",
            "originAddressId",
            "vendorId",
            "createdAt",
            "deletedAt"
          ],
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid",
              "description": "Unique product ID",
              "example": "d628c599-6282-4e35-b05b-4a6990e678fa"
            },
            "name": {
              "type": "string",
              "description": "Short UI description of the product",
              "example": "NOMAD 26"
            },
            "description": {
              "type": "string",
              "description": "Long UI description of the product",
              "example": "Outdoor activities were never so close"
            },
            "imageUrl": {
              "type": "string",
              "description": "URL of the product image",
              "example": "https://www.dimibike.com/images/products/large/cross-nomad-26-inch.jpg"
            },
            "price": {
              "type": "number",
              "description": "The price Buyer pays when buys the product",
              "example": 199
            },
            "basePrice": {
              "type": "number",
              "description": "The price Seller pays to produce the product",
              "example": 75
            },
            "props": {
              "type": "object",
              "description": "JSON list of product specific props, such as color, size, volume, km/h, etc.",
              "example": {
                "color": "Red"
              }
            },
            "totalInStock": {
              "type": "number",
              "description": "How many of these products we have in the store",
              "example": 11
            },
            "originAddressId": {
              "type": "string",
              "format": "uuid",
              "description": "Reference to the `AddressModel`",
              "example": "d628c599-6282-4e35-b05b-4a6990e678fa"
            },
            "vendorId": {
              "type": "string",
              "format": "uuid",
              "description": "Reference to the `VendorModel`",
              "example": "d628c599-6282-4e35-b05b-4a6990e678fa"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time",
              "readOnly": true,
              "description": "DateTime indicating when that product was created",
              "example": "2020-04-30T00:00:00.000Z"
            },
            "deletedAt": {
              "type": "string",
              "format": "date-time",
              "readOnly": true,
              "description": "DateTime indicating when that product was marked as deleted",
              "example": "2021-04-30T00:00:00.000Z"
            }
          }
        }
      ]
    },
...
    "UserModel": {
      "description": "User can be the administrator that oversee trade.\nUser can also manage stores.\n",
      "allOf": [
        {
          "type": "object",
          "required": [
            "id",
            "name",
            "email",
            "hashedPassword",
            "status",
            "createdAt",
            "deletedAt"
          ],
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid",
              "description": "Unique user ID",
              "example": "d628c599-6282-4e35-b05b-4a6990e678fa"
            },
            "name": {
              "type": "string",
              "description": "User name",
              "example": "John Smith"
            },
            "email": {
              "type": "string",
              "description": "User email",
              "example": "john@smith.com"
            },
            "hashedPassword": {
              "type": "string",
              "description": "User password after hashing",
              "example": "$argon2i$v=19$m=4096,t=3,p=1$e11PTDcTo+38dPL5QkNDYw$uVU+YbfAsp6PWQWAEKuDkj/u5PtULEEQAWFvreJwkDg"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time",
              "readOnly": true,
              "description": "DateTime indicating when that user was created",
              "example": "2020-04-30T00:00:00.000Z"
            },
            "deletedAt": {
              "type": "string",
              "format": "date-time",
              "readOnly": true,
              "description": "DateTime indicating when that user was marked as deleted",
              "example": "2021-04-30T00:00:00.000Z"
            }
          }
        }
      ]
    },
...
  },
  "responses": {
...
    "AppNotFound": {
      "description": "Some unexpected issue occurred and crashed the app.\n",
      "schema": {
        "$ref": "#/definitions/ResponseError"
      },
      "examples": {
        "application/json": {
          "id": "a34f4558-1b88-47cd-8d63-ae4eaf038bd7",
          "code": "INTERNAL_SERVER_ERROR",
          "message": "Internal server error"
        }
      }
    }
  },
...
}